Walking Stride Pattern Segmentation from Raw Accelerometry Data via ADEPT

Segment walking stride pattern from a raw accelerometry data time-series (x,y,z) via Adaptive Empirical Pattern Transformation (ADEPT). Default algorithm parameters are optimized for a wrist-worn sensor and were evaluated with data collected in the free-living environment.

segmentWalking(
  xyz,
  xyz.fs,
  template,
  sim_MIN = 0.85,
  dur_MIN = 0.8,
  dur_MAX = 1.4,
  ptp_r_MIN = 0.2,
  ptp_r_MAX = 2,
  vmc_r_MIN = 0.05,
  vmc_r_MAX = 0.5,
  mean_abs_diff_med_p_MAX = 0.5,
  mean_abs_diff_med_t_MAX = 0.2,
  mean_abs_diff_dur_MAX = 0.2,
  compute.template.idx = FALSE,
  run.parallel = FALSE,
  run.parallel.cores = 1
)

Arguments

xyz
A numeric matrix (or data frame) of n × 3 dimension. Three-dimensional raw accelerometry data time-series; acceleration measurements (x,y,z) collected along three orthogonal axes by the sensor's accelerometer.

xyz.fs
A numeric scalar. Frequency at which a time-series (x,y,z) is collected, expressed in a number of observations per second.

template
A list of numeric vectors, or a numeric vector. Distinct pattern template(s) of walking stride.

sim_MIN
numeric scalar. Minimum value of correlation between pattern template(s) and (r_t)_t vector magnitude of accelerometry data. Default used is 0.85.

dur_MIN
A numeric scalar. Minimum value of a stride duration allowed to be identified. Expressed in seconds. Default used is 0.8.

dur_MAX
A numeric scalar. Maximum value of a stride duration allowed to be identified. Expressed in seconds. Default used is 1.4.

ptp_r_MIN
A numeric scalar. Minimum value of "peak to peak" difference in (r_t)_t vector magnitude data of a stride. Default used is 0.2.

ptp_r_MAX
A numeric scalar. Maximum value of "peak to peak" difference in (r_t)_t vector magnitude data of a stride. Default used is 2.0

vmc_r_MIN
A numeric scalar. Minimum value of VMC in (r_t)_t vector magnitude data of a stride. Default used is 0.05.

vmc_r_MAX
A numeric scalar. Maximum value of VMC in (r_t)_t vector magnitude data of a stride. Default used is 0.5.

mean_abs_diff_med_p_MAX
A numeric scalar. Maximum value of MAD* of Azimuth (az_)_t median for 3 subsequent valid strides. Here, MAD* stands for mean of 2 absolute differences between 3 subsequent values. Default used is 0.5.

mean_abs_diff_med_t_MAX
A numeric scalar. Maximum value of MAD* of Elevation (el_)_t median for 3 subsequent valid strides. Here, MAD* stands for mean of 2 absolute differences between 3 subsequent values. Default used is 0.2.

mean_abs_diff_dur_MAX
A numeric scalar. Maximum value of MAD* of duration time for 3 subsequent valid strides. Here, MAD* stands for mean of 2 absolute differences between 3 subsequent values. Default used is 0.2.

compute.template.idx
A logical scalar. Whether or not to compute and return information about which of the provided pattern templates yielded a similarity matrix value that corresponds to an identified pattern occurrence. Setting to TRUE may increase computation time. Default is FALSE.

run.parallel
A logical scalar. Whether or not to use parallel execution in the algorithm with parallel package. Default is FALSE. DOES NOT WORK ON WINDOWS.

run.parallel.cores
An integer scalar. The number of cores to use for parallel execution. Defaults to 1L (no parallel). DOES NOT WORK ON WINDOWS.

Value

A data.frame with segmentation results. Each row describes one identified pattern occurrence:

tau_i - row index of xyz where pattern starts,

T_i - pattern duration, expressed in xyz vector length,

sim_i - similarity between a pattern and best-fit template; see segmentPattern for details,

template_i - if compute.template.idx equals TRUE: index of a template best matched to x; if compute.template.idx equals FALSE: NA,

is_walking_i -- 1 if a pattern is identified as walking stride; 0 otherwise.

Examples

library(adeptdata)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(adept)
xyz <-
  adeptdata::acc_walking_IU %>%
  filter(loc_id == "left_wrist", subj_id == "id86237981") %>%
  arrange(time_s) %>%
  select(v1 = x, v2 = y, v3 = z) %>%
  as.matrix()
# define raw accelerometry data sample frequency
xyz.fs <- 100

# define template list based on predefined templates
template_mat <- adeptdata::stride_template$left_wrist[[3]]
template <- list(
  template_mat[1, ],
  template_mat[2, ],
  template_mat[3, ]
)

# run walking segmentation
# (parallel supported, except for Windows; see run.parallel, run.parallel.cores args)
segmentWalking(xyz, xyz.fs, template)
#>     tau_i T_i     sim_i template_i is_walking_i
#> 1      32 108 0.9400755         NA            1
#> 2     139 111 0.8869175         NA            1
#> 3     249 109 0.9582450         NA            1
#> 4     357 104 0.9761546         NA            1
#> 5     460 104 0.9120968         NA            1
#> 6     563 107 0.9830195         NA            1
#> 7     669 107 0.9336347         NA            1
#> 8     775 107 0.9656279         NA            1
#> 9     881 109 0.9482314         NA            1
#> 10    989 104 0.9267459         NA            1
#> 11   1092 106 0.9602521         NA            1
#> 12   1197 109 0.9640643         NA            1
#> 13   1305 105 0.9618296         NA            1
#> 14   1409 104 0.9649911         NA            1
#> 15   1512 107 0.9560041         NA            1
#> 16   1618 105 0.9544636         NA            1
#> 17   1722 107 0.9637198         NA            1
#> 18   1828 106 0.9389940         NA            1
#> 19   1933 107 0.9648289         NA            1
#> 20   2039 105 0.9271943         NA            1
#> 21   2143 103 0.9669317         NA            1
#> 22   2245 108 0.9529903         NA            1
#> 23   2352 105 0.9388985         NA            1
#> 24   2456 103 0.9464128         NA            1
#> 25   2558 104 0.9625126         NA            1
#> 26   2661 103 0.9367948         NA            1
#> 27   2763 107 0.9590811         NA            1
#> 28   2869 107 0.9368955         NA            1
#> 29   2975 105 0.9508432         NA            1
#> 30   3079 101 0.9498385         NA            1
#> 31   3179 104 0.9522629         NA            1
#> 32   3282 107 0.9572315         NA            1
#> 33   3388 104 0.9522560         NA            1
#> 34   3491 106 0.9376362         NA            1
#> 35   3596 104 0.9664938         NA            1
#> 36   3699 105 0.9480016         NA            1
#> 37   3803 107 0.9479293         NA            1
#> 38   3909 105 0.9374259         NA            1
#> 39   4013 106 0.9644621         NA            1
#> 40   4118 106 0.9470178         NA            1
#> 41   4223 106 0.9425431         NA            1
#> 42   4328 104 0.9615261         NA            1
#> 43   4431 108 0.9551369         NA            1
#> 44   4538 106 0.9063153         NA            1
#> 45   4643 105 0.9644864         NA            1
#> 46   4747 107 0.9587655         NA            1
#> 47   4853 105 0.9371393         NA            1
#> 48   4957 103 0.9567233         NA            1
#> 49   5059 107 0.9680661         NA            1
#> 50   5165 107 0.9137281         NA            1
#> 51   5271 103 0.9575932         NA            1
#> 52   5373 107 0.9255886         NA            1
#> 53   5479 103 0.9615926         NA            1
#> 54   5581 106 0.9445123         NA            1
#> 55   5686 107 0.9554444         NA            1
#> 56   5792 104 0.9446203         NA            1
#> 57   5895 106 0.9346105         NA            1
#> 58   6000 103 0.9543212         NA            1
#> 59   6102 104 0.9461296         NA            1
#> 60   6205 106 0.9446571         NA            1
#> 61   6310 106 0.9504907         NA            1
#> 62   6415 108 0.9045167         NA            1
#> 63   6522  50 0.8119191         NA            0
#> 64   6571  56 0.6373326         NA            0
#> 65   6626 108 0.9389299         NA            1
#> 66   6733 103 0.9459896         NA            1
#> 67   6835 105 0.9558689         NA            1
#> 68   6939 105 0.9498884         NA            1
#> 69   7043 105 0.9582598         NA            1
#> 70   7147 105 0.9090522         NA            1
#> 71   7251 109 0.9606347         NA            1
#> 72   7359  99 0.8841097         NA            1
#> 73   7457 108 0.9650038         NA            1
#> 74   7564 104 0.9408641         NA            1
#> 75   7667 106 0.9586689         NA            1
#> 76   7772 102 0.9751309         NA            1
#> 77   7873 104 0.9470959         NA            1
#> 78   7976 104 0.9446294         NA            1
#> 79   8079 108 0.9525965         NA            1
#> 80   8186 103 0.9376855         NA            1
#> 81   8288 106 0.9535873         NA            1
#> 82   8393 102 0.9490769         NA            1
#> 83   8494 106 0.9256479         NA            1
#> 84   8599 105 0.9529229         NA            1
#> 85   8703 105 0.9425423         NA            1
#> 86   8807 104 0.9523595         NA            1
#> 87   8910 109 0.8230816         NA            0
#> 88   9018 105 0.9273690         NA            1
#> 89   9122 108 0.9315156         NA            1
#> 90   9229 103 0.9516589         NA            1
#> 91   9331 105 0.9146543         NA            1
#> 92   9435 105 0.9492184         NA            1
#> 93   9539 106 0.9451186         NA            1
#> 94   9644 102 0.9600201         NA            1
#> 95   9745 104 0.9106170         NA            1
#> 96   9848 105 0.9545347         NA            1
#> 97   9952  68 0.8761321         NA            0
#> 98  10057 105 0.9627304         NA            1
#> 99  10161 105 0.8815573         NA            1
#> 100 10265 117 0.9752817         NA            1
#> 101 10381  55 0.8403984         NA            0
#> 102 10472 104 0.9839812         NA            1
#> 103 10575 105 0.9670718         NA            1
#> 104 10679 103 0.8997071         NA            1
#> 105 10781 105 0.9820263         NA            1
#> 106 10885 106 0.9785955         NA            1
#> 107 10990 118 0.9220943         NA            1
#> 108 11107  93 0.9823024         NA            1
#> 109 11199 104 0.9221772         NA            1
#> 110 11302 108 0.9659211         NA            1
#> 111 11409 101 0.9585366         NA            1
#> 112 11509 107 0.9426814         NA            1
#> 113 11615 102 0.9206717         NA            1
#> 114 11716 101 0.9585659         NA            1
#> 115 11816 104 0.9538150         NA            1
#> 116 11919 106 0.8992340         NA            1
#> 117 12024 104 0.9525865         NA            1
#> 118 12127 103 0.9501956         NA            1
#> 119 12229 104 0.9806201         NA            1
#> 120 12332 104 0.9370644         NA            1
#> 121 12435 106 0.9382929         NA            1
#> 122 12540 105 0.9500415         NA            1
#> 123 12644 107 0.9628410         NA            1
#> 124 12750 105 0.9259910         NA            1
#> 125 12854 105 0.9807395         NA            1
#> 126 12958 103 0.9201171         NA            1
#> 127 13060 104 0.9736825         NA            1
#> 128 13163 103 0.9224098         NA            1
#> 129 13265 106 0.9754389         NA            1
#> 130 13370 104 0.9378643         NA            1
#> 131 13473 104 0.9380072         NA            1
#> 132 13576 108 0.9632811         NA            1
#> 133 13683 101 0.9317862         NA            1
#> 134 13783 106 0.9631919         NA            1
#> 135 13888 104 0.9034926         NA            1
#> 136 13991 109 0.9627612         NA            1
#> 137 14099 115 0.9375491         NA            1
#> 138 14213  56 0.8138986         NA            0
#> 139 14303 119 0.9747014         NA            0
#> 140 14421 106 0.9871113         NA            0
#> 141 14526  54 0.8038975         NA            0
#> 142 14618 105 0.9771843         NA            1
#> 143 14722 106 0.9114328         NA            1
#> 144 14827 104 0.9806270         NA            1
#> 145 14930 106 0.9068570         NA            1
#> 146 15035 106 0.9779649         NA            1
#> 147 15140 119 0.9775478         NA            1
#> 148 15258  55 0.7745829         NA            0
#> 149 15349 109 0.9703114         NA            1
#> 150 15457 113 0.8850693         NA            1
#> 151 15569 105 0.9812574         NA            1
#> 152 15673  93 0.9839535         NA            1
#> 153 15765 105 0.9726018         NA            1
#> 154 15869 105 0.9699617         NA            1
#> 155 15973 116 0.9552217         NA            1
#> 156 16088  56 0.8275288         NA            0
#> 157 16178 119 0.9738896         NA            1
#> 158 16296  94 0.9791999         NA            1
#> 159 16389 105 0.9273083         NA            1
#> 160 16493 106 0.9797275         NA            1
#> 161 16598 104 0.9499324         NA            1
#> 162 16701 106 0.8892281         NA            1
#> 163 16806 119 0.9822439         NA            1
#> 164 16924  93 0.9861907         NA            1
#> 165 17016 105 0.9430805         NA            1
#> 166 17120 106 0.9837561         NA            1
#> 167 17225 106 0.9272653         NA            1
#> 168 17330 120 0.9853681         NA            1
#> 169 17449  55 0.7453137         NA            0
#> 170 17542 108 0.9764402         NA            1
#> 171 17649 117 0.9579690         NA            1
#> 172 17765 102 0.9681750         NA            1
#> 173 17866  59 0.8249438         NA            0
#> 174 17962 116 0.9853900         NA            1
#> 175 18077  95 0.9774183         NA            1
#> 176 18171 104 0.9226323         NA            1
#> 177 18274 107 0.9829307         NA            1
#> 178 18380 106 0.9344850         NA            1
#> 179 18485 116 0.9863423         NA            1
#> 180 18600  55 0.8017060         NA            0
#> 181 18693 105 0.9844053         NA            1
#> 182 18797 118 0.9455927         NA            1
#> 183 18914 106 0.9737100         NA            1
#> 184 19019  97 0.8104342         NA            0
#> 185 19115 101 0.9781634         NA            1
#> 186 19215 109 0.9749692         NA            1
#> 187 19323 105 0.9253470         NA            1
#> 188 19427 106 0.9767969         NA            1
#> 189 19532 108 0.9126964         NA            1
#> 190 19639 106 0.9706312         NA            1
#> 191 19744 105 0.9221751         NA            1
#> 192 19848 117 0.9344927         NA            1
#> 193 19964  94 0.7661782         NA            0
#> 194 20057 107 0.9436316         NA            1
#> 195 20163 110 0.9280728         NA            1
#> 196 20272 108 0.9204715         NA            1
#> 197 20379 109 0.8320289         NA            0
#> 198 20487 112 0.9398095         NA            0
#> 199 20598  50 0.4477888         NA            0